refactor: move useRefToScrollToElseOptions into RadioSubPage
